Day 16
Initially was just brute force + memoization, but now is smarter
(calculate distances and compute the score when jumping directly to
next valve). Part 2 simply simulates 2 "rounds" of part 1 one after the
other (e.g. try to release as many valves for time=26, then do it again
with the current state)